First implementation of the javascript mcaro library extending AudioSwitcher functionality
Libraries are a to be considered as a javascript namespace with a set of functions.. eg...

AudioSwitcher.getAudioDevices(1)

Implemented JavaScriptAudioDevice, which is a Javascript friendly wrapper over an AudioDevice, it exposes a limited set of functionality availabel to js.

Core library, where helpful macro functions will be implemented. At the moment it has Core.sleep()